Update Index.cshtml
This change fixes a problem where newly added titles are added twice when Send is clicked; once by the SignalR processing and once on the done event for the Post. By calling the addBook function you avoid this because there is a check that only adds new books (not preexisting books).